Fix tests

+++ database/migrations/2018_04_12_000000_create_tenants_table.php
@@ -29,6 +29,9 @@ class CreateTenantsTable extends Migration
      */
     public function down()
     {
+        if(\App::environment('testing') && config('database.default') == 'sqlite') {
+            return;
+        }
         Schema::dropIfExists('tenants');
     }
 }
