+++ database/migrations/2019_06_06_115129_add_app_slug_column_to_engine_registries_table.php
@@ -25,6 +25,10 @@ class AddAppSlugColumnToEngineRegistriesTable extends Migration
      */
     public function down()
     {
+        if(\App::environment('testing') && config('database.default') == 'sqlite') {
+            return;
+        }
+
         Schema::table('engine_registries', function (Blueprint $table) {
             $table->dropColumn('slug');
         });
